% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pretest_ambig}
\alias{pretest_ambig}
\title{Pre-test calibration data}
\format{A data frame with 3,224 observations of 10 variables
\describe{
\item{trial}{Trial number}
\item{movie}{The stimulus file from this trial}
\item{aud}{Test stimulus}
\item{response}{Factor with levels "b" and "d"}
\item{time}{Reaction time, in seconds}
\item{source}{Name of the raw data file this observations came from}
\item{session_time}{The date and time of this session, based on the timestamp
  on the source data file}
\item{subject}{The subject number}
\item{response_raw}{Character of response key from raw data file.  Anything
  other than "Z" and "M" are coded as NA in response}
}}
\source{
Vroomen, J., van Linden, S., de Gelder, B., & Bertelson, P. (2007).
  Visual recalibration and selective adaptation in auditory-visual speech
  perception: Contrasting build-up courses.  Neuropsychologia, 45(3), 572–7.
  \url{https://doi.org/10.1016/j.neuropsychologia.2006.01.031}
}
\usage{
pretest_ambig
}
\description{
A second pre-test was done for some subjects, using only the three maximally
ambiguous stimuli that were later tested during adaptation.
}
\seealso{
\link{\code{pretest}} for the full continuum pretest data.
}
\keyword{datasets}
